backend/L2/requirements.txt
boto3==1.34.69
python-dotenv==1.0.1
redis==5.0.1
faker==22.6.0
pandas==2.2.1
numpy==1.26.4 